Refactored most of the AcsListener code out of the Program class and in to AcsListener
Fixes #62

Moved most of the AcsListener "worker" code out of the Program class and in to the AcsListener class.   Added a Start() method that is invoked to get an AcsListener to start working.
