+++ AcsListener/AcsListener/AcsListener.csproj
@@ -96,6 +96,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AcsListener.cs" />
     <Compile Include="AcsListenerConfigItems.cs" />
     <Compile Include="AcspAnnounceException.cs" />
     <Compile Include="AcspAnnounceRequest.cs" />
